// src/interfaces/senateInterfaces.ts
// Generated by https://quicktype.io

export interface SenateResponse {
  status: string;
  copyright: string;
  results: Result[];
}

export interface Result {
  congress: string;
  chamber: string;
  num_results: number;
  offset: number;
  members: Member[];
}

export interface Member {
  id: string;
  title: Title;
  short_title: ShortTitle;
  api_uri: string;
  first_name: string;
  middle_name: null | string;
  last_name: string;
  suffix: null | string;
  date_of_birth: string;
  gender: Gender;
  party: Party;
  leadership_role: null | string;
  twitter_account: null | string;
  facebook_account: null | string;
  youtube_account: null | string;
  govtrack_id: string;
  cspan_id: null | string;
  votesmart_id: null | string;
  icpsr_id: null | string;
  crp_id: string;
  google_entity_id: string;
  fec_candidate_id: string;
  url: string;
  rss_url: null | string;
  contact_form: null | string;
  in_office: boolean;
  cook_pvi: null;
  dw_nominate: number | null;
  ideal_point: null;
  seniority: string;
  next_election: string;
  total_votes: number;
  missed_votes: number;
  total_present: number;
  last_updated: LastUpdated;
  ocd_id: string;
  office: string;
  phone: string;
  fax: null | string;
  state: string;
  senate_class: string;
  state_rank: StateRank;
  lis_id: string;
  missed_votes_pct: number;
  votes_with_party_pct: number;
  votes_against_party_pct: number;
}

export enum Gender {
  F = "F",
  M = "M",
}

export enum LastUpdated {
  The202012301901140500 = "2020-12-30 19:01:14 -0500",
  The202012301901180500 = "2020-12-30 19:01:18 -0500",
  The202012301901190500 = "2020-12-30 19:01:19 -0500",
  The202012310808330500 = "2020-12-31 08:08:33 -0500",
  The202101021444020500 = "2021-01-02 14:44:02 -0500",
  The202101031140380500 = "2021-01-03 11:40:38 -0500",
  The202101031140400500 = "2021-01-03 11:40:40 -0500",
}

export enum Party {
  D = "D",
  ID = "ID",
  R = "R",
}

export enum ShortTitle {
  Sen = "Sen.",
}

export enum StateRank {
  Empty = "",
  Junior = "junior",
  Senior = "senior",
}

export enum Title {
  Senator1StClass = "Senator, 1st Class",
  Senator2NdClass = "Senator, 2nd Class",
  Senator3RDClass = "Senator, 3rd Class",
}